# sql: initialize virtual table descriptors concurrently; skip some vtable validations