See individual commits.

```
$ benchdiff --old 4348689e5d8 --new 959ef1f5c66 ./pkg/server -r BenchmarkTestServerStartup -c 10 -b -d 5s
old:  4348689 Merge #142011
new:  959ef1f sql: don't validate virtual tables upon creation u
args: benchdiff "--old" "4348689e5d8" "--new" "959ef1f5c66" "./pkg/server" "-r" "BenchmarkTestServerStartup" "-c" "10" "-b" "-d" "5s"

name                                   old time/op    new time/op    delta
TestServerStartup/SystemTenantOnly-24     388ms ± 4%     353ms ± 7%  -9.02%  (p=0.000 n=10+10)
TestServerStartup/ExternalTenant-24       389ms ± 4%     360ms ± 6%  -7.42%  (p=0.000 n=10+10)
TestServerStartup/SharedTenant-24         385ms ± 5%     363ms ± 6%  -5.75%  (p=0.000 n=9+10)

name                                   old alloc/op   new alloc/op   delta
TestServerStartup/SharedTenant-24         132MB ± 0%     132MB ± 0%  +0.21%  (p=0.003 n=10+10)
TestServerStartup/ExternalTenant-24       132MB ± 0%     132MB ± 0%  +0.28%  (p=0.001 n=10+10)
TestServerStartup/SystemTenantOnly-24     132MB ± 0%     132MB ± 0%  +0.39%  (p=0.000 n=9+9)

name                                   old allocs/op  new allocs/op  delta
TestServerStartup/SharedTenant-24          797k ± 0%      798k ± 0%  +0.16%  (p=0.000 n=10+10)
TestServerStartup/ExternalTenant-24        797k ± 0%      798k ± 0%  +0.19%  (p=0.000 n=10+10)
TestServerStartup/SystemTenantOnly-24      797k ± 0%      798k ± 0%  +0.21%  (p=0.000 n=10+10)
```

Epic: None